+++ src/StateSmith/Runner/ExceptionPrinter.cs
@@ -77,6 +77,12 @@ public class ExceptionPrinter
         return sb.ToString();
     }
 
+    public static string BuildReasonsString(string reasons)
+    {
+        string str = "Reason(s): " + reasons.ReplaceLineEndings("\n           ");
+        return str;
+    }
+
     public string? TryBuildingCustomExceptionDetails(System.Exception ex)
     {
         switch (ex)
@@ -84,14 +90,15 @@ public class ExceptionPrinter
             case DiagramEdgeParseException parseException:
                 {
                     DiagramEdge edge = parseException.edge;
+                    string reasons = ex.Message;
+                    reasons = BuildReasonsString(reasons);
                     string fromString = VertexPathDescriber.Describe(parseException.sourceVertex);
                     string toString = VertexPathDescriber.Describe(parseException.targetVertex);
-                    string reasons = ex.Message.ReplaceLineEndings("\n           ");
                     string message = $@"Failed parsing diagram edge
 from: {fromString}
 to:   {toString}
 Edge label: `{edge.label}`
-Reason(s): {reasons}
+{reasons}
 Edge diagram id: {edge.id}
 ";
                     return message;
@@ -119,7 +126,7 @@ Edge diagram id: {edge.id}
                 string fromString = VertexPathDescriber.Describe(vertex);
 
 message += $@"
-    Vertex
+Vertex:
     Path: {fromString}
     Diagram Id: {vertex.DiagramId}
     Children count: {vertex.Children.Count}
@@ -139,7 +146,7 @@ message += $@"
                     string toString = VertexPathDescriber.Describe(behavior.TransitionTarget);
 
                     message += $@"
-    Behavior
+Behavior:
     Owning vertex: {fromString}
     Target vertex: {toString}
     Order: {behavior.GetOrderString()}