+++ src/StateSmith/Runner/IConsolePrinter.cs
@@ -2,6 +2,7 @@ namespace StateSmith.Runner;
 
 public interface IConsolePrinter
 {
+    void SetLineEnding(string lineEnding);
     void OutputStageMessage(string message) => WriteLine("StateSmith Runner - " + message);
     void WriteErrorLine(string message);
     void WriteLine();