+++ src/StateSmithTest/Input/Expansions/ExpandingVisitorTest.cs
@@ -2,9 +2,12 @@ using Xunit;
 using FluentAssertions;
 using StateSmith.Input.Expansions;
 using StateSmith.Input.Antlr4;
+using StateSmithTest.Output.UserConfig;
+using StateSmithTest.Output;
+using StateSmith.SmGraph.Validation;
+using System.Text;
 
 namespace StateSmithTest.Input.Expansions;
-
 
 public class ExpandingVisitorTest
 {